Implemented Main functionality

+++ app/src/main/java/com/popda/weatherforcast/data/entity/WeatherForecast.kt
@@ -26,6 +26,6 @@ data class WeatherForecast (
 	@SerializedName("wind_deg") val wind_deg : Int,
 	@SerializedName("weather") val weather : List<Weather>,
 	@SerializedName("clouds") val clouds : Int,
-	@SerializedName("pop") val pop : Int,
+	@SerializedName("pop") val pop : Double,
 	@SerializedName("uvi") val uvi : Double
 )
